feat: [#8] int 및 Integer 값 Long으로 수정

+++ src/main/java/org/rf/rfserver/domain/Schedule.java
@@ -19,14 +19,14 @@ public class Schedule {
     private String scheduleName;
     private LocalDateTime datetime;
     private String location;
-    private int participantCount;
-    private Integer alert;
+    private Long participantCount;
+    private Long alert;
 
     @ManyToOne(fetch = LAZY)
     private Party party;
 
     @Builder
-    public Schedule(Long id, String scheduleName, LocalDateTime datetime, String location, int participantCount, Integer alert, Party party){
+    public Schedule(Long id, String scheduleName, LocalDateTime datetime, String location, Long participantCount, Long alert, Party party){
         this.id = id;
         this.scheduleName = scheduleName;
         this.datetime = datetime;
@@ -36,7 +36,7 @@ public class Schedule {
         this.party = party;
     }
 
-    public Schedule updateSchedule(String scheduleName, LocalDateTime datetime, String location, Integer alert){
+    public Schedule updateSchedule(String scheduleName, LocalDateTime datetime, String location, Long alert){
         this.scheduleName = scheduleName ==  null ? this.scheduleName : scheduleName;
         this.datetime = datetime == null ? this.datetime : datetime;
         this.location = location == null ? this.location : location;
